# web-rally/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rally Tascas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rally/v1/rally/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "View Rally Settings",
        "description": "View rally settings",
        "operationId": "view_rally_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallySettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Rally Settings",
        "description": "Update global rally configuration (admin only).\nArgs:\n    settings_in: New settings values\n\nReturns:\n    Updated rally settings\n\nRaises:\n    403: If user is not authorized\n    400: If validation fails",
        "operationId": "update_rally_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RallySettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallySettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/rally/v1/rally/settings/banner": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Upload Rally Banner",
        "description": "Upload the event banner image to Cloudflare R2 (admin only).",
        "operationId": "upload_rally_banner",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_rally_banner"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallySettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          },
          "403": {
            "description": "Not authorized"
          },
          "503": {
            "description": "R2 storage not configured or upload failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rally/v1/rally/settings/logo": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Upload Rally Logo",
        "description": "Upload the event logo image to Cloudflare R2 (admin only).",
        "operationId": "upload_rally_logo",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_rally_logo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallySettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          },
          "403": {
            "description": "Not authorized"
          },
          "503": {
            "description": "R2 storage not configured or upload failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rally/v1/rally/settings/favicon": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Upload Rally Favicon",
        "description": "Upload the browser-tab favicon to Cloudflare R2 (admin only). Accepts png/ico/svg.",
        "operationId": "upload_rally_favicon",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_rally_favicon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallySettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          },
          "403": {
            "description": "Not authorized"
          },
          "503": {
            "description": "R2 storage not configured or upload failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_rally_banner": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_upload_rally_banner"
      },
      "Body_upload_rally_favicon": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_upload_rally_favicon"
      },
      "Body_upload_rally_logo": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_upload_rally_logo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeSection": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "visible": {
            "type": "boolean",
            "title": "Visible",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "HomeSection"
      },
...
      "RallySettingsResponse": {
        "properties": {
          "max_teams": {
            "type": "integer",
            "title": "Max Teams"
          },
          "max_members_per_team": {
            "type": "integer",
            "title": "Max Members Per Team"
          },
          "enable_versus": {
            "type": "boolean",
            "title": "Enable Versus"
          },
          "penalty_per_puke": {
            "type": "integer",
            "title": "Penalty Per Puke"
          },
          "penalty_per_not_drinking": {
            "type": "integer",
            "title": "Penalty Per Not Drinking"
          },
          "bonus_per_extra_shot": {
            "type": "integer",
            "title": "Bonus Per Extra Shot"
          },
          "max_extra_shots_per_member": {
            "type": "integer",
            "title": "Max Extra Shots Per Member"
          },
          "checkpoint_order_matters": {
            "type": "boolean",
            "title": "Checkpoint Order Matters"
          },
          "enable_staff_scoring": {
            "type": "boolean",
            "title": "Enable Staff Scoring"
          },
          "show_live_leaderboard": {
            "type": "boolean",
            "title": "Show Live Leaderboard"
          },
          "show_team_details": {
            "type": "boolean",
            "title": "Show Team Details"
          },
          "show_checkpoint_map": {
            "type": "boolean",
            "title": "Show Checkpoint Map"
          },
          "participant_view_enabled": {
            "type": "boolean",
            "title": "Participant View Enabled"
          },
          "show_route_mode": {
            "type": "string",
            "title": "Show Route Mode"
          },
          "show_score_mode": {
            "type": "string",
            "title": "Show Score Mode"
          },
          "rally_theme": {
            "type": "string",
            "title": "Rally Theme"
          },
          "event_name": {
            "type": "string",
            "title": "Event Name",
            "default": "Rally Tascas"
          },
          "event_subtitle": {
            "type": "string",
            "title": "Event Subtitle",
            "default": ""
          },
          "accent_color": {
            "type": "string",
            "title": "Accent Color",
            "default": ""
          },
          "public_access_enabled": {
            "type": "boolean",
            "title": "Public Access Enabled"
          },
          "allow_staff_registration": {
            "type": "boolean",
            "title": "Allow Staff Registration",
            "default": false
          },
          "allow_photo_as_team_photo": {
            "type": "boolean",
            "title": "Allow Photo As Team Photo",
            "default": false
          },
          "guide_mode_enabled": {
            "type": "boolean",
            "title": "Guide Mode Enabled",
            "default": false
          },
          "guide_mode_active": {
            "type": "boolean",
            "title": "Guide Mode Active",
            "default": false
          },
          "badges_enabled": {
            "type": "boolean",
            "title": "Badges Enabled",
            "default": true
          },
          "home_layout": {
            "items": {
              "$ref": "#/components/schemas/HomeSection"
            },
            "type": "array",
            "title": "Home Layout",
            "default": [
              {
                "key": "home_hero",
                "visible": true
              },
              {
                "key": "rally_marquee",
                "visible": true
              },
              {
                "key": "event_stats",
                "visible": true
              },
              {
                "key": "live_top5",
                "visible": true
              },
              {
                "key": "how_it_works",
                "visible": true
              },
              {
                "key": "postos_preview",
                "visible": true
              },
              {
                "key": "home_bottom_banner",
                "visible": true
              }
            ]
          },
          "ticker_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ticker Items",
            "default": [
              "RALLY TASCAS 2026",
              "NEI",
              "EQUIPAS",
              "POSTOS",
              "PONTUA\u00c7\u00c3O"
            ]
          },
          "rally_start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rally Start Time"
          },
          "rally_end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rally End Time"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "default": "rally_tascas"
          },
          "banner_url": {
            "type": "string",
            "title": "Banner Url",
            "default": ""
          },
          "logo_url": {
            "type": "string",
            "title": "Logo Url",
            "default": ""
          },
          "favicon_url": {
            "type": "string",
            "title": "Favicon Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "max_teams",
          "max_members_per_team",
          "enable_versus",
          "penalty_per_puke",
          "penalty_per_not_drinking",
          "bonus_per_extra_shot",
          "max_extra_shots_per_member",
          "checkpoint_order_matters",
          "enable_staff_scoring",
          "show_live_leaderboard",
          "show_team_details",
          "show_checkpoint_map",
          "participant_view_enabled",
          "show_route_mode",
          "show_score_mode",
          "rally_theme",
          "public_access_enabled"
        ],
        "title": "RallySettingsResponse"
      },
      "RallySettingsUpdate": {
        "properties": {
          "max_teams": {
            "type": "integer",
            "title": "Max Teams"
          },
          "max_members_per_team": {
            "type": "integer",
            "title": "Max Members Per Team"
          },
          "enable_versus": {
            "type": "boolean",
            "title": "Enable Versus"
          },
          "penalty_per_puke": {
            "type": "integer",
            "title": "Penalty Per Puke"
          },
          "penalty_per_not_drinking": {
            "type": "integer",
            "title": "Penalty Per Not Drinking"
          },
          "bonus_per_extra_shot": {
            "type": "integer",
            "title": "Bonus Per Extra Shot"
          },
          "max_extra_shots_per_member": {
            "type": "integer",
            "title": "Max Extra Shots Per Member"
          },
          "checkpoint_order_matters": {
            "type": "boolean",
            "title": "Checkpoint Order Matters"
          },
          "enable_staff_scoring": {
            "type": "boolean",
            "title": "Enable Staff Scoring"
          },
          "show_live_leaderboard": {
            "type": "boolean",
            "title": "Show Live Leaderboard"
          },
          "show_team_details": {
            "type": "boolean",
            "title": "Show Team Details"
          },
          "show_checkpoint_map": {
            "type": "boolean",
            "title": "Show Checkpoint Map"
          },
          "participant_view_enabled": {
            "type": "boolean",
            "title": "Participant View Enabled"
          },
          "show_route_mode": {
            "type": "string",
            "title": "Show Route Mode"
          },
          "show_score_mode": {
            "type": "string",
            "title": "Show Score Mode"
          },
          "rally_theme": {
            "type": "string",
            "title": "Rally Theme"
          },
          "event_name": {
            "type": "string",
            "title": "Event Name",
            "default": "Rally Tascas"
          },
          "event_subtitle": {
            "type": "string",
            "title": "Event Subtitle",
            "default": ""
          },
          "accent_color": {
            "type": "string",
            "title": "Accent Color",
            "default": ""
          },
          "public_access_enabled": {
            "type": "boolean",
            "title": "Public Access Enabled"
          },
          "allow_staff_registration": {
            "type": "boolean",
            "title": "Allow Staff Registration",
            "default": false
          },
          "allow_photo_as_team_photo": {
            "type": "boolean",
            "title": "Allow Photo As Team Photo",
            "default": false
          },
          "guide_mode_enabled": {
            "type": "boolean",
            "title": "Guide Mode Enabled",
            "default": false
          },
          "guide_mode_active": {
            "type": "boolean",
            "title": "Guide Mode Active",
            "default": false
          },
          "badges_enabled": {
            "type": "boolean",
            "title": "Badges Enabled",
            "default": true
          },
          "home_layout": {
            "items": {
              "$ref": "#/components/schemas/HomeSection"
            },
            "type": "array",
            "title": "Home Layout",
            "default": [
              {
                "key": "home_hero",
                "visible": true
              },
              {
                "key": "rally_marquee",
                "visible": true
              },
              {
                "key": "event_stats",
                "visible": true
              },
              {
                "key": "live_top5",
                "visible": true
              },
              {
                "key": "how_it_works",
                "visible": true
              },
              {
                "key": "postos_preview",
                "visible": true
              },
              {
                "key": "home_bottom_banner",
                "visible": true
              }
            ]
          },
          "ticker_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ticker Items",
            "default": [
              "RALLY TASCAS 2026",
              "NEI",
              "EQUIPAS",
              "POSTOS",
              "PONTUA\u00c7\u00c3O"
            ]
          }
        },
        "type": "object",
        "required": [
          "max_teams",
          "max_members_per_team",
          "enable_versus",
          "penalty_per_puke",
          "penalty_per_not_drinking",
          "bonus_per_extra_shot",
          "max_extra_shots_per_member",
          "checkpoint_order_matters",
          "enable_staff_scoring",
          "show_live_leaderboard",
          "show_team_details",
          "show_checkpoint_map",
          "participant_view_enabled",
          "show_route_mode",
          "show_score_mode",
          "rally_theme",
          "public_access_enabled"
        ],
        "title": "RallySettingsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}